# Support internvl-chat for pytorch engine