--- a/lmdeploy/pytorch/config.py
+++ b/lmdeploy/pytorch/config.py
@@ -1,9 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from dataclasses import dataclass, field
-from typing import Any, Dict, List, Literal
+from typing import Any, Dict, List
 
 import torch
-from transformers import AutoModelForCausalLM
 
 
 def _update_torch_dtype(config: 'ModelConfig', default: str = 'float16'):
@@ -81,8 +80,7 @@ class ModelConfig:
     init_kwargs: Dict[str, Any] = field(default_factory=dict)
     model_arch: str = None
     unused_modules: List[str] = None
-    task_type: Literal['llm', 'vlm'] = 'llm'
-    auto_model_cls: Any = AutoModelForCausalLM
+    auto_model_cls: Any = None
 
     def get_head_size(self):
         """get head size."""
@@ -101,7 +99,6 @@ class ModelConfig:
     @classmethod
     def from_hf_config(cls, hf_config: Any, model_path: str = None):
         """from huggingface config."""
-        from lmdeploy.archs import check_vl_llm
         from lmdeploy.pytorch.configurations import AutoModelConfigBuilder
 
         model_config = AutoModelConfigBuilder.build(hf_config, model_path)
@@ -110,9 +107,6 @@ class ModelConfig:
         model_config.model_arch = model_arch
         # should after setting `hf_config` and `model_arch` attributes
         model_config = _update_torch_dtype(model_config)
-
-        if check_vl_llm(model_config.hf_config.to_dict()):
-            model_config.task_type = 'vlm'
 
         # update eos_token_id to list
         if isinstance(model_config.eos_token_id, int):