--- a/lmdeploy/pytorch/configurations/builder.py
+++ b/lmdeploy/pytorch/configurations/builder.py
@@ -49,4 +49,35 @@ class AutoModelConfigBuilder(ABC):
         cfg = valid_builder.build(hf_config, model_path)
         if cfg.hf_config is None:
             cfg.hf_config = hf_config
+
+        if cfg.auto_model_cls is None:
+            cfg.auto_model_cls = ProxyAutoModel()
         return cfg
+
+
+class ProxyAutoModel:
+    """wrapper of auto model class."""
+
+    def __init__(self, model_cls=None) -> None:
+        """init."""
+        if model_cls is None:
+            from transformers import AutoModelForCausalLM
+            model_cls = AutoModelForCausalLM
+        self.model_cls = model_cls
+
+    def from_config(self, *args, **kwargs):
+        """wrap from_config."""
+        if hasattr(self.model_cls, '_from_config'):
+            if 'trust_remote_code' in kwargs:
+                kwargs.pop('trust_remote_code')
+            return self.model_cls._from_config(*args, **kwargs)
+        else:
+            return self.model_cls.from_config(*args, **kwargs)
+
+    def from_pretrained(self, *args, **kwargs):
+        """wrap from_pretrained."""
+        if hasattr(self.model_cls,
+                   '_from_config') and 'trust_remote_code' in kwargs:
+            kwargs.pop('trust_remote_code')
+
+        return self.model_cls.from_pretrained(*args, **kwargs)