--- a/lmdeploy/pytorch/engine/engine_instance.py
+++ b/lmdeploy/pytorch/engine/engine_instance.py
@@ -479,6 +479,8 @@ class EngineInstance:
 
     def decode(self,
                input_ids,
+               input_embeddings: List[InputEmbeddingType] = None,
+               input_embedding_ranges: List[InputEmbeddingRangeType] = None,
                steps: List[int] = None,
                sequence_start: bool = True,
                sequence_end: bool = True,
@@ -496,19 +498,30 @@ class EngineInstance:
         logger.debug('Decoding logits.')
         batch_size = len(input_ids)
 
-        def __add_messages(session_ids, input_ids, adapter_names):
+        def __add_messages(session_ids, input_ids, adapter_names,
+                           input_embeddings, input_embedding_ranges):
             add_msgs = []
             sampling_param = SamplingParam(max_new_tokens=0)
-            for session_id, token_id, adapter_name in zip(
-                    session_ids, input_ids, adapter_names):
+            for (session_id, token_id, adapter_name, input_emb,
+                 input_ranges) in zip(session_ids, input_ids, adapter_names,
+                                      input_embeddings,
+                                      input_embedding_ranges):
                 if len(token_id) > self.max_input_len:
                     raise RuntimeError(
                         f'Expect input length<={self.max_input_len} '
                         f'but get {len(token_id)}')
+                cur_input_embeddings: List[InputEmbeddings] = None
+                if input_emb is not None and len(input_emb) > 0:
+                    assert len(input_emb) == len(input_ranges)
+                    cur_input_embeddings = [
+                        InputEmbeddings(emb, rg[0], rg[1])
+                        for emb, rg in zip(input_emb, input_ranges)
+                    ]
                 msg = dict(token_ids=token_id,
                            session_id=session_id,
                            sampling_param=sampling_param,
                            adapter_name=adapter_name,
+                           input_embeddings=cur_input_embeddings,
                            return_logits=True)
                 add_msgs.append(msg)
             req_types = [RequestType.ADD_MESSAGE] * batch_size
@@ -519,9 +532,17 @@ class EngineInstance:
         if steps is not None:
             assert batch_size == len(steps)
 
-        if adapter_names is None:
+        if adapter_names is not None:
+            assert len(adapter_names) == batch_size
+        else:
             adapter_names = [None] * batch_size
-        assert batch_size == len(adapter_names)
+
+        if input_embeddings is not None:
+            assert len(input_embeddings) == batch_size
+            assert len(input_embedding_ranges) == batch_size
+        else:
+            input_embeddings = [None] * batch_size
+            input_embedding_ranges = [None] * batch_size
 
         session_ids = tuple(range(batch_size))
         if sequence_start:
@@ -530,7 +551,8 @@ class EngineInstance:
                                      dict(session_id=sid))
                 self._try_add_session(sid)
 
-        req_ids = __add_messages(session_ids, input_ids, adapter_names)
+        req_ids = __add_messages(session_ids, input_ids, adapter_names,
+                                 input_embeddings, input_embedding_ranges)
         req_idx_map = dict(zip(req_ids, range(len(req_ids))))
 
         finish_count = batch_size