--- a/lmdeploy/pytorch/engine/model_agent.py
+++ b/lmdeploy/pytorch/engine/model_agent.py
@@ -652,13 +652,11 @@ class BaseModelAgent(AutoModelAgent):
         device = 'cuda'
         with LoadNoInit(), warnings.catch_warnings():
             warnings.simplefilter('ignore')
-            model_init_kwargs = dict(torch_dtype=torch_dtype,
-                                     device_map=device)
-            model_init_kwargs.update(self.model_config.init_kwargs)
-            if hasattr(self.model_config.auto_model_cls, 'from_config'):
-                model_init_kwargs['trust_remote_code'] = trust_remote_code
             hf_model = self.model_config.auto_model_cls.from_pretrained(
-                model_path, **model_init_kwargs)
+                model_path,
+                torch_dtype=torch_dtype,
+                device_map=device,
+                **self.model_config.init_kwargs)
             hf_model.eval()
             hf_model.config.use_cache = True
             # build for vlm model
@@ -815,18 +813,13 @@ def _tp_build_model(
         config = model_config.hf_config
         torch_dtype = model_config.dtype
         device_map = None
-        model_init_kwargs = dict(torch_dtype=torch_dtype)
-        model_init_kwargs.update(model_config.init_kwargs)
-        # deal with AutoModel or Non-AutoModel
-        if hasattr(model_config.auto_model_cls, 'from_config'):
-            model_init_kwargs['trust_remote_code'] = trust_remote_code
-            from_config_func = model_config.auto_model_cls.from_config
-        else:
-            from_config_func = model_config.auto_model_cls._from_config
-
         with init_empty_weights(), warnings.catch_warnings():
             warnings.simplefilter('ignore')
-            model = from_config_func(config, **model_init_kwargs)
+            model = model_config.auto_model_cls.from_config(
+                config,
+                torch_dtype=torch_dtype,
+                trust_remote_code=trust_remote_code,
+                **model_config.init_kwargs)
             # build for vlm model
             _remove_unused_modules(model, model_config)
             if rank == 0:
@@ -835,6 +828,7 @@ def _tp_build_model(
             if rank == 0:
                 # adapter would remove weight of linear.
                 device_map = _create_device_map(model, world_size, device_map)
+
         model.eval()
         model.config.use_cache = True