… docs/en/supported_models/supported_models.md
… docs/zh_cn/supported_models/supported_models.md
… lmdeploy/pytorch/models/module_map.py
… lmdeploy/pytorch/supported_models.py