@RunningLeon 2024-06-21
@grimoire 
## performance on llama3-8b
cmd
```
python benchmark/profile_throughput.py ./ShareGPT_V3_unfiltered_cleaned_split.json /nvme/shared_data/llama3/Meta-Llama-3-8B-Instruct --backend pytorch -n 3000
```

### main branch
```
concurrency: 256
elapsed_time: 199.272s

first token latency(s)(min, max, ave): 1.238, 11.655, 2.881
per-token latency(s) percentile(50, 75, 95, 99): [0.04, 0.042, 0.278, 0.34]

number of prompt tokens: 676779
number of completion tokens: 612685
token throughput (completion token): 3074.616 token/s
token throughput (prompt + completion token): 6470.873 token/s
RPS (request per second): 15.055 req/s
RPM (request per minute): 903.288 req/min
--------------------------------------------------
```
### this PR
```
concurrency: 256
elapsed_time: 205.748s

first token latency(s)(min, max, ave): 1.033, 11.620, 2.934
per-token latency(s) percentile(50, 75, 95, 99): [0.041, 0.044, 0.279, 0.343]

number of prompt tokens: 676779
number of completion tokens: 612685
token throughput (completion token): 2977.838 token/s
token throughput (prompt + completion token): 6267.192 token/s
RPS (request per second): 14.581 req/s
RPM (request per minute): 874.855 req/min
--------------------------------------------------
```