@lvhan028 2024-06-24
My test code is as follows, and the test command is `python examples/workspace/test_vl_pipeline.py /nvme/shared/InternVL-Chat-V1-5 --backend pytorch --cache_max_entry_count 0.5`
```python
from lmdeploy import pipeline, TurbomindEngineConfig, PytorchEngineConfig
from lmdeploy.vl import load_image
import fire
import time

def apply_turbomind(model_path, **kwargs):
    engine_config = TurbomindEngineConfig.from_dict({}, allow_none=True)

    for key, value in kwargs.items():
        if hasattr(TurbomindEngineConfig, key):
            setattr(engine_config, key, value)

    pipe = pipeline(model_path, backend_config=engine_config)

    return pipe


def apply_pytorch(model_path, **kwargs):
    engine_config = PytorchEngineConfig()
    for key, value in kwargs.items():
        print(key, value)
        if hasattr(PytorchEngineConfig, key):
            setattr(engine_config, key, value)
            
    pipe = pipeline(model_path, backend_config=engine_config)
    
    return pipe


def main(model_path, backend, **kwargs):
    start = time.perf_counter()
    if backend == 'turbomind':
        pipe = apply_turbomind(model_path, **kwargs)
    else:
        pipe = apply_pytorch(model_path, **kwargs)
    end = time.perf_counter()
    print(f'building pipeline cost: {end - start} s')
    
    input('press any key to continue')
    image = load_image('https://raw.githubusercontent.com/open-mmlab/mmdeploy/main/tests/data/tiger.jpeg')
    response = pipe(('describe this image', image))
    print(response)
    input('press any key to exit')
    
if __name__ == "__main__":
    fire.Fire(main)
    

```