@v3ss0n 2024-07-13
I had build using this PR , and when inference i got this error :

Mini-InternVL-Chat-4B-V1-5

AssertionError: turbomind does not support /home/nozander/.cache/huggingface/hub/models--OpenGVLab--Mini-InternVL-Chat-4B-V1-5/snapshots/920f7428f246dbb460a1d90c4a3ee9dc696158cc. Plz try pytorch engine instead.