@RunningLeon 2024-07-15
@v3ss0n hi, pls. use `lmdeploy check_env` to check your env info and post a sample code to reproduce.